+++ packages/css-color-parser/src/functions/hsl.ts
@@ -1,7 +1,8 @@
-import type { FunctionNode } from '@csstools/css-parser-algorithms';
-import { ColorData, SyntaxFlag } from '../color-data';
+import type { ColorData } from '../color-data';
 import type { ColorParser } from '../color-parser';
+import type { FunctionNode } from '@csstools/css-parser-algorithms';
 import { ColorNotation } from '../color-notation';
+import { SyntaxFlag } from '../color-data';
 import { normalize_legacy_HSL_ChannelValues, normalize_modern_HSL_ChannelValues } from './hsl-normalize-channel-values';
 import { threeChannelLegacySyntax } from './three-channel-legacy-syntax';
 import { threeChannelSpaceSeparated } from './three-channel-space-separated';