+++ packages/css-color-parser/src/functions/hwb.ts
@@ -1,6 +1,6 @@
-import type { FunctionNode } from '@csstools/css-parser-algorithms';
-import { ColorData } from '../color-data';
+import type { ColorData } from '../color-data';
 import type { ColorParser } from '../color-parser';
+import type { FunctionNode } from '@csstools/css-parser-algorithms';
 import { ColorNotation } from '../color-notation';
 import { normalize_HWB_ChannelValues } from './hwb-normalize-channel-values';
 import { threeChannelSpaceSeparated } from './three-channel-space-separated';