+++ packages/css-color-parser/src/functions/oklch.ts
@@ -2,8 +2,8 @@ import type { ColorData } from '../color-data';
 import type { ColorParser } from '../color-parser';
 import type { FunctionNode } from '@csstools/css-parser-algorithms';
 import { ColorNotation } from '../color-notation';
-import { threeChannelSpaceSeparated } from './three-channel-space-separated';
 import { normalize_OKLCH_ChannelValues } from './oklch-normalize-channel-values';
+import { threeChannelSpaceSeparated } from './three-channel-space-separated';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export function oklch(oklchNode: FunctionNode, colorParser: ColorParser): ColorData | false {