+++ packages/css-color-parser/src/functions/rgb-normalize-channel-values.ts
@@ -1,7 +1,9 @@
-import { CSSToken, NumberType, TokenNumber, TokenType } from '@csstools/css-tokenizer';
-import { ColorData, SyntaxFlag } from '../color-data';
-import { toLowerCaseAZ } from '../util/to-lower-case-a-z';
+import type { ColorData } from '../color-data';
+import type { TokenNumber } from '@csstools/css-tokenizer';
+import { CSSToken, NumberType, TokenType } from '@csstools/css-tokenizer';
+import { SyntaxFlag } from '../color-data';
 import { normalize } from './normalize';
+import { toLowerCaseAZ } from '../util/to-lower-case-a-z';
 
 export function normalize_legacy_sRGB_ChannelValues(tokens: Array<CSSToken>, colorData: ColorData): Array<TokenNumber> | false {
 	const result: Array<TokenNumber> = [];