+++ plugins/postcss-color-mix-function/.tape.mjs
@@ -1,0 +1,23 @@
+import { postcssTape } from '@csstools/postcss-tape';
+import plugin from '@csstools/postcss-color-mix-function';
+
+postcssTape(plugin)({
+	basic: {
+		message: "supports basic usage",
+	},
+	'basic:preserve-true': {
+		message: "supports basic usage with { preserve: true }",
+		options: {
+			preserve: true
+		}
+	},
+	'examples/example': {
+		message: 'minimal example',
+	},
+	'examples/example:preserve-true': {
+		message: 'minimal example',
+		options: {
+			preserve: true
+		}
+	},
+});