+++ plugins/postcss-oklab-function/src/index.ts
@@ -2,7 +2,7 @@ import postcssProgressiveCustomProperties from '@csstools/postcss-progressive-cu
 import type { Declaration } from 'postcss';
 import type { PluginCreator } from 'postcss';
 import { cloneTokens, tokenize } from '@csstools/css-tokenizer';
-import { color, serializeP3, serializeRGB, SyntaxFlag } from '@csstools/css-color-parser';
+import { color, colorDataTo, ColorNotation, serializeP3, serializeRGB, SyntaxFlag } from '@csstools/css-color-parser';
 import { hasFallback } from './has-fallback-decl';
 import { hasSupportsAtRuleAncestor } from './has-supports-at-rule-ancestor';
 import { isFunctionNode, parseCommaSeparatedListOfComponentValues, replaceComponentValues, stringify } from '@csstools/css-parser-algorithms';
@@ -74,6 +74,11 @@ const basePlugin: PluginCreator<basePluginOptions> = (opts?: basePluginOptions) 
 								return;
 							}
 
+							const srgb = colorDataTo(colorData, ColorNotation.RGB);
+							if (!srgb.channels.find((x) => x <= 0.00001 || x >= 0.99999)) {
+								return serializeRGB(colorData);
+							}
+
 							return serializeP3(colorData);
 						}
 					},
@@ -82,7 +87,7 @@ const basePlugin: PluginCreator<basePluginOptions> = (opts?: basePluginOptions) 
 
 			decl.cloneBefore({ value: modifiedRGB });
 
-			if (opts?.subFeatures.displayP3) {
+			if (opts?.subFeatures.displayP3 && modifiedP3 !== modifiedRGB) {
 				decl.cloneBefore({ value: modifiedP3 });
 			}